propagate error codes, and factorize configuration

+++ scripts/process_lhef.sh
@@ -1,11 +1,5 @@
 #!/bin/bash
 
-# configure the location of the pythia and Delphes executables
-MG_DIR=/data7/atlas/cshimmin/zplite/MG5_aMC
-
-PYTHIA_DIR=$MG_DIR/pythia-pgs/src
-DELPHES_DIR=/gdata/atlas/cshimmin/zplite/outputs/delphes
-
 LHAPATH=$PYTHIA_DIR/PDFsets
 
 # need to export this magic variable for pythia
@@ -23,8 +17,12 @@ fi
 if [ ! -e "unweighted_events.lhe" ]; then
 	if [ -e "unweighted_events.lhe.gz" ]; then
 		gunzip unweighted_events.lhe.gz
+		if [ $? -ne 0 ]; then
+			echo "Failed to unzip LHE!" >&2
+			exit 1
+		fi
 	else
-		echo "No events file found!"
+		echo "No events file found!" >&2
 		exit 1
 	fi
 fi
@@ -32,8 +30,20 @@ fi
 # run pythia
 $PYTHIA_DIR/pythia
 
+pythia_code=$?
+if [ $pythia_code -ne 0 ]; then
+	echo "Pythia returned error! Code=$pythia_code" >&2
+	exit $pythia_code
+fi
+
 # run delphes
 $DELPHES_DIR/DelphesSTDHEP ../Cards/delphes_card.dat delphes.root pythia_events.hep
+
+delphes_code=$?
+if [ $delphes_code -ne 0 ]; then
+	echo "Delphes returned error! Code=$delphes_code" >&2
+	exit $delphes_code
+fi
 
 # clean the hep file
 rm pythia_events.hep
